==> 01-CreditApprovalUCI_dataPrep.ipynb <==
In this notebook, we prepare the Credit Approval data set from the UCI Machine Learning Repository, to leave it more suitable for the demos of the recipes from chapter 3.

===========================================================================

To download the Credit Approval dataset from the UCI Machine Learning Repository visit [this website](http://archive.ics.uci.edu/ml/machine-learning-databases/credit-screening/) and click on **crx.data** and **crx.names** to download data and variable names. Save **crx.data** to the parent folder of your notebook folder.

**Citation:**

Dua, D. and Graff, C. (2019). UCI Machine Learning Repository [http://archive.ics.uci.edu/ml]. Irvine, CA: University of California, School of Information and Computer Science.

import random
import pandas as pd
import numpy as np

!dir

# load data
data = pd.read_csv('../crx.data', header=None)

# create variable names according to UCI Machine Learning
# Repo information
varnames = ['A'+str(s) for s in range(1,17)]

# add column names
data.columns = varnames

# replace ? by np.nan
data = data.replace('?', np.nan)

# re-cast some variables to the correct types 
data['A2'] = data['A2'].astype('float')
data['A14'] = data['A14'].astype('float')

# encode target to binary
data['A16'] = data['A16'].map({'+':1, '-':0})
print(data.dtypes)
data.head()

## add more missing values to random positions
# this will help with the demos of the recipes

random.seed(9001)

values = set([random.randint(0, len(data)) for p in range(0, 100)])
print(len(values))
for var in ['A3', 'A8', 'A9', 'A10']:
    data.loc[values, var] = np.nan
    
    
data.isnull().sum()

# save the data
data.to_csv('../creditApprovalUCI.csv', index=False)

data.head()

# find categorical variables
cat_cols = [c for c in data.columns if data[c].dtypes=='O']
data[cat_cols].head()

# find numerical variables
num_cols = [c for c in data.columns if data[c].dtypes!='O']
data[num_cols].head()
