==> 01-Titanic_dataPrep.ipynb <==
## Predicting Survival on the Titanic

### History
Perhaps one of the most infamous shipwrecks in history, the Titanic sank after colliding with an iceberg, killing 1502 out of 2224 people on board. Interestingly, by analysing the probability of survival based on few attributes like gender, age, and social status, we can make very accurate predictions on which passengers would survive. Some groups of people were more likely to survive than others, such as women, children, and the upper-class. Therefore, we can learn about the society priorities and privileges at the time.

### Dataset


### Download and Save

import pandas as pd
import numpy as np

data = pd.read_csv('https://www.openml.org/data/get_csv/16826755/phpMYEkMl')
data.head()

There are strings containing "?" which do not get considered as null values. We will need to replace these with `np.nan`

data = data.replace('?', np.nan)
data.isnull().sum()

def get_first_cabin(row):
    try:
        return row.split()[0]
    except:
        return np.nan 

data['cabin'] = data['cabin'].apply(get_first_cabin)

data.to_csv('../titanic.csv', index=False)
