==> 04.22-Missing-Indicator-Feature-Engine.ipynb <==
## Missing Indicator ==> Feature-Engine


### What is Feature-Engine

Feature-Engine is an open source python package that I created at the back of this course. 

- Feature-Engine includes all the feature engineering techniques described in the course
- Feature-Engine works like to Scikit-learn, so it is easy to learn
- Feature-Engine allows you to implement specific engineering steps to specific feature subsets
- Feature-Engine can be integrated with the Scikit-learn pipeline allowing for smooth model building
- 
**Feature-Engine allows you to design and store a feature engineering pipeline with bespoke procedures for different variable groups.**

-------------------------------------------------------------------
Feature-Engine can be installed via pip ==> pip install feature-engine

- Make sure you have installed feature-engine before running this notebook

For more information visit:
my website

## In this demo

We will use Feature-Engine to perform mean or median imputation using the Ames House Price Dataset.

- To download the dataset visit the lecture **Datasets** in **Section 1** of the course.

import pandas as pd
import numpy as np

import matplotlib.pyplot as plt

# to split the datasets
from sklearn.model_selection import train_test_split
from sklearn.pipeline import Pipeline

# from feature-engine
from feature_engine.imputation import (
    AddMissingIndicator,
    MeanMedianImputer,
    CategoricalImputer
)

# let's load the dataset with a selected group of variables

cols_to_use = [
    'BsmtQual', 'FireplaceQu', 'LotFrontage', 'MasVnrArea', 'GarageYrBlt',
    'SalePrice'
]

data = pd.read_csv('../houseprice.csv', usecols=cols_to_use)
data.head()

data.isnull().mean()

# let's separate into training and testing set

# first drop the target from the feature list
cols_to_use.remove('SalePrice')

X_train, X_test, y_train, y_test = train_test_split(data[cols_to_use],
                                                    data['SalePrice'],
                                                    test_size=0.3,
                                                    random_state=0)
X_train.shape, X_test.shape

### Feature-Engine's missing indicator selects all variables by default

# we call the imputer from feature-engine
# the argument how allows us to determine if we want
# to add missing indicators to all variables, or only to
# those that showed missing data in the train set

imputer = AddMissingIndicator(missing_only=True)

# we fit the imputer

imputer.fit(X_train)

# different from other transformers, here, the attribute
# variables shows the variables entered by the user, in this
# case None

imputer.variables

# this attribute stores the variables, numerical and categorical,
# if they showed missing data in the train set

imputer.variables_

# feature-engine returns a dataframe
# with the additional features

# no need to contatenate!!

tmp = imputer.transform(X_train)
tmp.head()

# let's check NA

tmp.isnull().mean()

## Feature-engine allows you to specify variable groups easily

# let's do it imputation but this time
# and let's select a few variables

imputer = AddMissingIndicator(variables=['BsmtQual', 'FireplaceQu', 'LotFrontage'])

imputer.fit(X_train)

# now the imputer uses only the variables we indicated

imputer.variables

# missing indicators will be added for the following variables
# in case that these are different from the ones passed by the user

# remember that with the argument how set to 'missing_only' the imputer
# will learn and store the variables if they show NA in the train dataset

imputer.variables_

# feature-engine returns a dataframe
# with the additional features

# no need to contatenate!!

tmp = imputer.transform(X_train)

tmp.head()

## Feature-engine can be used with the Scikit-learn pipeline

# let's check the percentage of NA in each variable

X_train.isnull().mean()

X_train.head()

These are the steps we will concatenate

- Add Missing Indicator to all variables
- Median Imputation to numerical variables
- Frequent category imputation to categorical variables

pipe = Pipeline([
    ('missing_ind', AddMissingIndicator()),

    ('imputer_mode', CategoricalImputer(
        imputation_method='frequent', variables=['FireplaceQu', 'BsmtQual'])),

    ('imputer_median', MeanMedianImputer(imputation_method='median',
                                         variables=['LotFrontage', 'MasVnrArea', 'GarageYrBlt'])),
])

# fit the pipe
pipe.fit(X_train)

# inspect the separate steps
pipe.named_steps['missing_ind'].variables_

pipe.named_steps['imputer_mode'].imputer_dict_

pipe.named_steps['imputer_median'].imputer_dict_

# let's transform the data with the pipeline

# this pipeline will:
#- add the missing indicators
#- fill na in the original variables
# leaving the dataset ready to use for ML

tmp = pipe.transform(X_train)

# let's check null values are gone
tmp.isnull().mean()

tmp.shape

For more details on the AddMissingIndicator visit the documentation:

https://feature-engine.readthedocs.io/en/latest/imputers/AddMissingIndicator.html
