==> 10.07-Further-reading-on-scaling.ipynb <==
## Additional reading resources

### Sklearn docs

- [Compare the effect of different scalers on data with outliers](http://scikit-learn.org/stable/auto_examples/preprocessing/plot_all_scaling.html#sphx-glr-auto-examples-preprocessing-plot-all-scaling-py)
- [Sklearn documentation on preprocessing and scaling](https://scikit-learn.org/stable/modules/preprocessing.html)

### Other blogs and articles
- [About Feature Scaling and Normalization](https://sebastianraschka.com/Articles/2014_about_feature_scaling.html)
- [Normalization vs Standardization — Quantitative analysis](https://towardsdatascience.com/normalization-vs-standardization-quantitative-analysis-a91e8a79cebf)
- [Z scores: Use & misuse](https://influentialpoints.com/Training/z_scores_use_and_misuse.htm)
- [L0 Norm, L1 Norm, L2 Norm & L-Infinity Norm](https://medium.com/@montjoile/l0-norm-l1-norm-l2-norm-l-infinity-norm-7a7d18a4f40c)
