==> 12.01_Engineering_dates.ipynb <==
## Engineering Dates

Date variables are special type of categorical variable. By their own nature, date variables will contain a multitude of different labels, each one corresponding to a specific date and sometimes time. Date variables, when preprocessed properly can highly enrich a dataset. For example, from a date variable we can extract:

- Week of the year
- Month
- Quarter
- Semester
- Year
- Day (number)
- Day of the week
- Is Weekend?
- Time differences in years, months, days, hrs, etc.

Date variables should not be used as categorical variables when building a machine learning model. Not only because they have a multitude of categories, but also because when we actually use the model to score a new observation, this observation will most likely be in the future, an therefore its date label, might be different from the ones contained in the training set and therefore the ones used to train the machine learning algorithm.


## In this demo: Peer to peer lending (Finance)

In this demo, we will use the toy data from the peer-o-peer finance company to extract different features from datetime variables.

import pandas as pd
import numpy as np

import datetime

# let's load the dataset with a few selected columns

use_cols = ['date_issued', 'date_last_payment']

data = pd.read_csv('../loan.csv', usecols=use_cols)

data.head()

# now let's parse the dates, currently cast as strings, into datetime format

data['issue_dt'] = pd.to_datetime(data['date_issued'])
data['last_pymnt_dt'] = pd.to_datetime(data['date_last_payment'])

data[['date_issued','issue_dt','date_last_payment', 'last_pymnt_dt']].head()

### Extract week of the year

# Extracting week of year from date, varies from 1 to 52

data['issue_dt_week'] = data['issue_dt'].dt.week

data[['issue_dt', 'issue_dt_week']].head()

data['issue_dt_week'].unique()

### Extract month

# Extracting month from date - 1 to 12

data['issue_dt_month'] = data['issue_dt'].dt.month

data[['issue_dt', 'issue_dt_month']].head()

data['issue_dt_month'].unique()

### Extract quarter

# Extract quarter from date variable - 1 to 4

data['issue_dt_quarter'] = data['issue_dt'].dt.quarter

data[['issue_dt', 'issue_dt_quarter']].head()

data['issue_dt_quarter'].unique()

### Extract semester

# We could also extract semester

data['issue_dt_semester'] = np.where(data['issue_dt_quarter'].isin([1,2]), 1, 2)

data.head()

data['issue_dt_semester'].unique()

###  Extract year

# extract year 

data['issue_dt_year'] = data['issue_dt'].dt.year

data[['issue_dt', 'issue_dt_year']].head()

data['issue_dt_year'].unique()

### Extract days, in various formats

# day - numeric from 1-31

data['issue_dt_day'] = data['issue_dt'].dt.day

data[['issue_dt', 'issue_dt_day']].head()

data['issue_dt_day'].unique()

# day of the week - from 0 to 6

data['issue_dt_dayofweek'] = data['issue_dt'].dt.dayofweek

data[['issue_dt', 'issue_dt_dayofweek']].head()

data['issue_dt_dayofweek'].unique()

# day of the week - name

data['issue_dt_dayofweek'] = data['issue_dt'].dt.day_name()

data[['issue_dt', 'issue_dt_dayofweek']].head()

data['issue_dt_dayofweek'].unique()

# was the application done on the weekend?

data['issue_dt_is_weekend'] = np.where(data['issue_dt_dayofweek'].isin(['Sunday', 'Saturday']), 1,0)

data[['issue_dt', 'issue_dt_dayofweek','issue_dt_is_weekend']].head()

data['issue_dt_is_weekend'].unique()

### Extract Time elapsed between dates

# perhaps more interestingly, extract the date difference between 2 dates

data['issue_dt'] - data['last_pymnt_dt']

# same as above capturing just the time difference

(data['last_pymnt_dt'] - data['issue_dt']).dt.days.head()

# calculate number of months passed between 2 dates

data['months_passed'] = (data['last_pymnt_dt'] - data['issue_dt']) / np.timedelta64(1, 'M')
data['months_passed'] = np.round(data['months_passed'],0)

data[['last_pymnt_dt', 'issue_dt','months_passed']].head()

# or the time difference to today

(datetime.datetime.today() - data['issue_dt']).head()

(datetime.datetime.today() - data['issue_dt']).unique()
